LayoutTests/fast/dom/Window/window-property-invalid-characters-ignored-expected.txt

Tests that invalid characters are ignored after a window property value.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS w.document.width is 123
PASS w.document.height is 123
PASS w !== window is true
PASS successfullyParsed is true

TEST COMPLETE
